BAVL-235 pulling out court email creation from facade into a factory.
